requirements.txt
Flask==3.0.0
plivo==4.44.0
python-dotenv==1.0.1
gunicorn==21.2.0
